Fix border issues & update miniCal.
Update the month selector on the top content bar to send users to the first of the month when they're interested in switching months. This removes the edge case previously outlined with the miniCal being on the 31st and going back to the 1st.

Also, changed the borders of events to a lighter white instead of making the border disappear.